DBちょっと修正

+++ db/migrate/20231119002946_create_lines.rb
@@ -3,7 +3,7 @@ class CreateLines < ActiveRecord::Migration[7.1]
     create_table :lines do |t|
       t.integer :from_node_id, null: false, index: true
       t.integer :to_node_id, null: false, index: true
-      t.integer :distance, null: false
+      t.float :distance, null: false
 
       t.timestamps
     end
